MetadataWidget : New module to hold the MetadataWidgets
By making the MetadataWidgets public, we should be able to make custom
Plug widgets possible.
