+++ python/GafferUI/UIEditor.py
@@ -39,7 +39,6 @@ import functools
 import types
 import re
 import collections
-import os
 import imath
 import inspect
 
@@ -47,6 +46,7 @@ import IECore
 
 import Gaffer
 import GafferUI
+import MetadataWidget
 
 ## The UIEditor class allows the user to edit the interfaces for nodes.
 class UIEditor( GafferUI.NodeSetEditor ) :
@@ -83,7 +83,7 @@ class UIEditor( GafferUI.NodeSetEditor ) :
 					_Label( "Description", parenting = { "verticalAlignment" : GafferUI.ListContainer.VerticalAlignment.Top } )
 
 					self.__nodeMetadataWidgets.append(
-						_MultiLineStringMetadataWidget( key = "description" )
+						MetadataWidget.MultiLineStringMetadataWidget( key = "description" )
 					)
 
 				with _Row() :
@@ -91,7 +91,7 @@ class UIEditor( GafferUI.NodeSetEditor ) :
 					_Label( "Documentation URL" )
 
 					self.__nodeMetadataWidgets.append(
-						_StringMetadataWidget( key = "documentation:url" )
+						MetadataWidget.StringMetadataWidget( key = "documentation:url" )
 					)
 
 				with _Row() :
@@ -99,7 +99,7 @@ class UIEditor( GafferUI.NodeSetEditor ) :
 					_Label( "Color" )
 
 					self.__nodeMetadataWidgets.append(
-						_ColorSwatchMetadataWidget( key = "nodeGadget:color" )
+						MetadataWidget.ColorSwatchMetadataWidget( key = "nodeGadget:color" )
 					)
 
 				with _Row() as self.__iconRow :
@@ -107,7 +107,7 @@ class UIEditor( GafferUI.NodeSetEditor ) :
 					_Label( "Icon" )
 
 					self.__nodeMetadataWidgets.append(
-						_FileSystemPathMetadataWidget( key = "icon" )
+						MetadataWidget.FileSystemPathMetadataWidget( key = "icon" )
 					)
 
 			# Plugs tab
@@ -320,325 +320,6 @@ class _Row( GafferUI.ListContainer ) :
 	def __init__( self, *args, **kw ) :
 
 		GafferUI.ListContainer.__init__( self, GafferUI.ListContainer.Orientation.Horizontal, spacing = 4, *args, **kw )
-
-##########################################################################
-# MetadataValueWidgets. These display metadata values, allowing the user
-# to edit them.
-##########################################################################
-
-class _MetadataWidget( GafferUI.Widget ) :
-
-	def __init__( self, topLevelWidget, key, target = None, **kw ) :
-
-		GafferUI.Widget.__init__( self, topLevelWidget, **kw )
-
-		self.__key = key
-		self.__target = None
-
-		self.setTarget( target )
-
-	def setTarget( self, target ) :
-
-		assert( isinstance( target, ( Gaffer.Node, Gaffer.Plug, type( None ) ) ) )
-
-		self.__target = target
-		self.setEnabled( self.__target is not None )
-
-		if isinstance( self.__target, Gaffer.Node ) :
-			self.__metadataChangedConnection = Gaffer.Metadata.nodeValueChangedSignal().connect(
-				Gaffer.WeakMethod( self.__nodeMetadataChanged )
-			)
-		elif isinstance( self.__target, Gaffer.Plug ) :
-			self.__metadataChangedConnection = Gaffer.Metadata.plugValueChangedSignal().connect(
-				Gaffer.WeakMethod( self.__plugMetadataChanged )
-			)
-		else :
-			self.__metadataChangedConnection = None
-
-		self.__update()
-
-	def getTarget( self ) :
-
-		return self.__target
-
-	def setKey( self, key ) :
-
-		if key == self.__key :
-			return
-
-		self.__key = key
-		self.__update()
-
-	def getKey( self, key ) :
-
-		return self.__key
-
-	## Must be implemented in derived classes to update
-	# the widget from the value.
-	def _updateFromValue( self, value ) :
-
-		raise NotImplementedError
-
-	## Must be called by derived classes to update
-	# the Metadata value when the widget value changes.
-	def _updateFromWidget( self, value ) :
-
-		if self.__target is None :
-			return
-
-		with Gaffer.UndoScope( self.__target.ancestor( Gaffer.ScriptNode ) ) :
-			Gaffer.Metadata.registerValue( self.__target, self.__key, value )
-
-	## May be called by derived classes to deregister the
-	# metadata value.
-	def _deregisterValue( self ) :
-
-		if self.__target is None :
-			return
-
-		with Gaffer.UndoScope( self.__target.ancestor( Gaffer.ScriptNode ) ) :
-			Gaffer.Metadata.deregisterValue( self.__target, self.__key )
-
-	def __update( self ) :
-
-		if self.__target is None :
-			self._updateFromValue( None )
-			return
-
-		v = Gaffer.Metadata.value( self.__target, self.__key )
-		if v is None :
-			k = self.__fallbackKey( self.__key )
-			if k is not None :
-				v = Gaffer.Metadata.value( self.__target, k )
-
-		self._updateFromValue( v )
-
-	def __nodeMetadataChanged( self, nodeTypeId, key, node ) :
-
-		if self.__key != key :
-			return
-		if node is not None and not node.isSame( self.__target ) :
-			return
-		if not self.__target.isInstanceOf( nodeTypeId ) :
-			return
-
-		self.__update()
-
-	def __plugMetadataChanged( self, nodeTypeId, plugPath, key, plug ) :
-
-		if self.__key != key :
-			return
-
-		if Gaffer.MetadataAlgo.affectedByChange( self.__target, nodeTypeId, plugPath, plug ) :
-			self.__update()
-
-	@staticmethod
-	def __fallbackKey( k ) :
-
-		for oldPrefix, newPrefix in [
-			( "pathPlugValueWidget:", "path:" ),
-			( "fileSystemPathPlugValueWidget:", "fileSystemPath:" ),
-		] :
-			if k.startswith( newPrefix ) :
-				return k.replace( newPrefix, oldPrefix )
-
-		return None
-
-class _BoolMetadataWidget( _MetadataWidget ) :
-
-	def __init__( self, key, target = None, **kw ) :
-
-		self.__boolWidget = GafferUI.BoolWidget()
-		_MetadataWidget.__init__( self, self.__boolWidget, key, target, **kw )
-
-		self.__boolWidget.stateChangedSignal().connect(
-			Gaffer.WeakMethod( self.__stateChanged ), scoped = False
-		)
-
-	def _updateFromValue( self, value ) :
-
-		self.__boolWidget.setState( value if value is not None else False )
-
-	def __stateChanged( self, *unused ) :
-
-		self._updateFromWidget( self.__boolWidget.getState() )
-
-class _StringMetadataWidget( _MetadataWidget ) :
-
-	def __init__( self, key, target = None, acceptEmptyString = True, **kw ) :
-
-		self.__textWidget = GafferUI.TextWidget()
-		_MetadataWidget.__init__( self, self.__textWidget, key, target, **kw )
-
-		self.__acceptEmptyString = acceptEmptyString
-
-		self.__textWidget.editingFinishedSignal().connect(
-			Gaffer.WeakMethod( self.__editingFinished ), scoped = False
-		)
-
-	def textWidget( self ) :
-
-		return self.__textWidget
-
-	def _updateFromValue( self, value ) :
-
-		self.__textWidget.setText( str( value ) if value is not None else "" )
-
-	def __editingFinished( self, *unused ) :
-
-		text = self.__textWidget.getText()
-		if text or self.__acceptEmptyString :
-			self._updateFromWidget( text )
-		else :
-			self._deregisterValue()
-
-class _MultiLineStringMetadataWidget( _MetadataWidget ) :
-
-	def __init__( self, key, target = None, role = GafferUI.MultiLineTextWidget.Role.Text, **kw ) :
-
-		self.__textWidget = GafferUI.MultiLineTextWidget( role = role )
-		_MetadataWidget.__init__( self, self.__textWidget, key, target, **kw )
-
-		self.__textWidget.editingFinishedSignal().connect(
-			Gaffer.WeakMethod( self.__editingFinished ), scoped = False
-		)
-
-	def textWidget( self ) :
-
-		return self.__textWidget
-
-	def _updateFromValue( self, value ) :
-
-		self.__textWidget.setText( value if value is not None else "" )
-
-	def __editingFinished( self, *unused ) :
-
-		self._updateFromWidget( self.__textWidget.getText() )
-
-class _ColorSwatchMetadataWidget( _MetadataWidget ) :
-
-	def __init__( self, key, target = None, **kw ) :
-
-		self.__swatch = GafferUI.ColorSwatch( useDisplayTransform = False )
-
-		_MetadataWidget.__init__( self, self.__swatch, key, target, **kw )
-
-		self.__swatch._qtWidget().setFixedHeight( 18 )
-		self.__swatch._qtWidget().setMaximumWidth( 40 )
-		self.__value = None
-
-		self.__swatch.buttonReleaseSignal().connect( Gaffer.WeakMethod( self.__buttonRelease ), scoped = False )
-
-	def _updateFromValue( self, value ) :
-
-		if value is not None :
-			self.__swatch.setColor( value )
-		else :
-			self.__swatch.setColor( imath.Color4f( 0, 0, 0, 0 ) )
-
-		self.__value = value
-
-	def __buttonRelease( self, swatch, event ) :
-
-		if event.button != event.Buttons.Left :
-			return False
-
-		color = self.__value if self.__value is not None else imath.Color3f( 1 )
-		dialogue = GafferUI.ColorChooserDialogue( color = color, useDisplayTransform = False )
-		color = dialogue.waitForColor( parentWindow = self.ancestor( GafferUI.Window ) )
-
-		if color is not None :
-			self._updateFromWidget( color )
-
-class _MenuMetadataWidget( _MetadataWidget ) :
-
-	def __init__( self, key, labelsAndValues, target = None, **kw ) :
-
-		self.__menuButton = GafferUI.MenuButton(
-			menu = GafferUI.Menu( Gaffer.WeakMethod( self.__menuDefinition ) )
-		)
-
-		self.__labelsAndValues = labelsAndValues
-		self.__currentValue = None
-
-		_MetadataWidget.__init__( self, self.__menuButton, key, target, **kw )
-
-	def _updateFromValue( self, value ) :
-
-		self.__currentValue = value
-
-		buttonText = str( value )
-		for label, value in self.__labelsAndValues :
-			if value == self.__currentValue :
-				buttonText = label
-				break
-
-		self.__menuButton.setText( buttonText )
-
-	def __menuDefinition( self ) :
-
-		result = IECore.MenuDefinition()
-		for label, value in self.__labelsAndValues :
-			result.append(
-				"/" + label,
-				{
-					"command" : functools.partial( Gaffer.WeakMethod( self.__setValue ), value = value ),
-					"checkBox" : value == self.__currentValue
-				}
-			)
-
-		return result
-
-	def __setValue( self, unused, value ) :
-
-		self._updateFromWidget( value )
-
-class _FileSystemPathMetadataWidget( _MetadataWidget ) :
-
-	def __init__( self, key, target = None, acceptEmptyString = True, **kw ) :
-
-		self.__row = GafferUI.ListContainer( GafferUI.ListContainer.Orientation.Horizontal, spacing = 4 )
-
-		self.__path = Gaffer.FileSystemPath()
-		self.__pathWidget = GafferUI.PathWidget( self.__path )
-
-		_MetadataWidget.__init__( self, self.__row, key, target, **kw )
-
-		self.__row.append( self.__pathWidget )
-
-		button = GafferUI.Button( image = "pathChooser.png", hasFrame=False )
-		button.clickedSignal().connect( Gaffer.WeakMethod( self.__buttonClicked ), scoped = False )
-		self.__row.append( button )
-
-		self.__acceptEmptyString = acceptEmptyString
-
-		self.__pathWidget.editingFinishedSignal().connect(
-			Gaffer.WeakMethod( self.__editingFinished ), scoped = False
-		)
-
-	def _updateFromValue( self, value ) :
-
-		self.__path.setFromString( str( value ) if value is not None else "" )
-
-	def __editingFinished( self, *unused ) :
-
-		text = str( self.__path )
-		if text or self.__acceptEmptyString :
-			self._updateFromWidget( text )
-		else :
-			self._deregisterValue()
-
-	def __buttonClicked( self, widget ) :
-
-		path = str( self.__path )
-		path = path if os.path.exists( path ) else os.path.expanduser( "~" )
-
-		dialogue = GafferUI.PathChooserDialogue( Gaffer.FileSystemPath( path ) )
-		chosenPath = dialogue.waitForPath( parentWindow = self.ancestor( GafferUI.Window ) )
-
-		if chosenPath is not None :
-			self.__path.setFromString( str( chosenPath ) )
-			self.__editingFinished()
 
 ##########################################################################
 # Hierarchical representation of a plug layout, suitable for manipulating
@@ -1571,12 +1252,12 @@ class _PlugEditor( GafferUI.Widget ) :
 			with _Row() :
 
 				_Label( "Label" )
-				self.__metadataWidgets["label"] = _StringMetadataWidget( key = "label", acceptEmptyString = False )
+				self.__metadataWidgets["label"] = MetadataWidget.StringMetadataWidget( key = "label", acceptEmptyString = False )
 
 			with _Row() :
 
 				_Label( "Description", parenting = { "verticalAlignment" : GafferUI.ListContainer.VerticalAlignment.Top } )
-				self.__metadataWidgets["description"] = _MultiLineStringMetadataWidget( key = "description" )
+				self.__metadataWidgets["description"] = MetadataWidget.MultiLineStringMetadataWidget( key = "description" )
 				self.__metadataWidgets["description"].textWidget().setFixedLineHeight( 10 )
 
 			with _Row() :
@@ -1618,7 +1299,7 @@ class _PlugEditor( GafferUI.Widget ) :
 					with _Row() :
 
 						_Label( "Position" )
-						self.__metadataWidgets["noduleLayout:section"] = _MenuMetadataWidget(
+						self.__metadataWidgets["noduleLayout:section"] = MetadataWidget.MenuMetadataWidget(
 							key = "noduleLayout:section",
 							labelsAndValues = [
 								( "Default", None ),
@@ -1632,12 +1313,12 @@ class _PlugEditor( GafferUI.Widget ) :
 					with _Row() :
 
 						_Label( "Color" )
-						self.__metadataWidgets["nodule:color"] = _ColorSwatchMetadataWidget( key = "nodule:color" )
+						self.__metadataWidgets["nodule:color"] = MetadataWidget.ColorSwatchMetadataWidget( key = "nodule:color" )
 
 					with _Row() :
 
 						_Label( "Connection Color" )
-						self.__metadataWidgets["connectionGadget:color"] = _ColorSwatchMetadataWidget( key = "connectionGadget:color" )
+						self.__metadataWidgets["connectionGadget:color"] = MetadataWidget.ColorSwatchMetadataWidget( key = "connectionGadget:color" )
 
 			GafferUI.Spacer( imath.V2i( 0 ), parenting = { "expand" : True } )
 
@@ -1819,22 +1500,22 @@ class _PlugEditor( GafferUI.Widget ) :
 
 	__MetadataDefinition = collections.namedtuple( "MetadataDefinition", ( "key", "label", "metadataWidgetType", "plugValueWidgetType" ) )
 	__metadataDefinitions = (
-		__MetadataDefinition( "fileSystemPath:extensions", "File Extensions", _StringMetadataWidget, "GafferUI.FileSystemPath*PlugValueWidget" ),
-		__MetadataDefinition( "path:bookmarks", "Bookmarks Category", _StringMetadataWidget, "GafferUI.FileSystemPath*PlugValueWidget" ),
-		__MetadataDefinition( "path:valid", "File Must Exist", _BoolMetadataWidget, "GafferUI.FileSystemPath*PlugValueWidget" ),
-		__MetadataDefinition( "path:leaf", "No Directories", _BoolMetadataWidget, "GafferUI.FileSystemPath*PlugValueWidget" ),
-		__MetadataDefinition( "fileSystemPath:includeSequences", "Allow sequences", _BoolMetadataWidget, "GafferUI.FileSystemPath*PlugValueWidget" ),
+		__MetadataDefinition( "fileSystemPath:extensions", "File Extensions", MetadataWidget.StringMetadataWidget, "GafferUI.FileSystemPath*PlugValueWidget" ),
+		__MetadataDefinition( "path:bookmarks", "Bookmarks Category", MetadataWidget.StringMetadataWidget, "GafferUI.FileSystemPath*PlugValueWidget" ),
+		__MetadataDefinition( "path:valid", "File Must Exist", MetadataWidget.BoolMetadataWidget, "GafferUI.FileSystemPath*PlugValueWidget" ),
+		__MetadataDefinition( "path:leaf", "No Directories", MetadataWidget.BoolMetadataWidget, "GafferUI.FileSystemPath*PlugValueWidget" ),
+		__MetadataDefinition( "fileSystemPath:includeSequences", "Allow sequences", MetadataWidget.BoolMetadataWidget, "GafferUI.FileSystemPath*PlugValueWidget" ),
 		# Note that includeSequenceFrameRange is primarily used by GafferCortex.
 		# Think twice before using it elsewhere	as it may not exist in the future.
-		__MetadataDefinition( "fileSystemPath:includeSequenceFrameRange", "Sequences include frame range", _BoolMetadataWidget, "GafferUI.FileSystemPath*PlugValueWidget" ),
+		__MetadataDefinition( "fileSystemPath:includeSequenceFrameRange", "Sequences include frame range", MetadataWidget.BoolMetadataWidget, "GafferUI.FileSystemPath*PlugValueWidget" ),
 		__MetadataDefinition(
 			"buttonPlugValueWidget:clicked",
 			"Button Click Code",
-			lambda key : _MultiLineStringMetadataWidget( key, role = GafferUI.MultiLineTextWidget.Role.Code ),
+			lambda key : MetadataWidget.MultiLineStringMetadataWidget( key, role = GafferUI.MultiLineTextWidget.Role.Code ),
 			"GafferUI.ButtonPlugValueWidget"
 		),
-		__MetadataDefinition( "layout:accessory", "Inline", _BoolMetadataWidget,  "GafferUI.ButtonPlugValueWidget" ),
-		__MetadataDefinition( "divider", "Divider", _BoolMetadataWidget,  "*" ),
+		__MetadataDefinition( "layout:accessory", "Inline", MetadataWidget.BoolMetadataWidget,  "GafferUI.ButtonPlugValueWidget" ),
+		__MetadataDefinition( "divider", "Divider", MetadataWidget.BoolMetadataWidget,  "*" ),
 
 	)
 
@@ -1870,7 +1551,7 @@ class _SectionEditor( GafferUI.Widget ) :
 
 				_Label( "Summary", parenting = { "verticalAlignment" : GafferUI.ListContainer.VerticalAlignment.Top } )
 
-				self.__summaryMetadataWidget = _MultiLineStringMetadataWidget( key = "" )
+				self.__summaryMetadataWidget = MetadataWidget.MultiLineStringMetadataWidget( key = "" )
 
 		self.__section = ""
 		self.__plugParent = None
